requirements.txt
kagglehub
pandas
matplotlib
seaborn
python-dotenv
langchain[all]
transformers
gradio
ipywidgets
ipykernel